fix(hogql): cast varAssignment tuple index for gcc, bump npm parser
`visitAssignmentTarget` assigned the result of `stoll` (a `long long`)
straight into a `Json` value. macOS clang accepts it, but gcc on CI
rejects it as an ambiguous `operator=` overload, which broke the wheel
build and every Django test shard (the suite rebuilds the C++ parser
when `common/hogql_parser/` changes). Cast to `int64_t`, matching the
existing `ColumnExprTupleAccess` visitor.

Also bumps the `@posthog/hogql-parser` npm package to 1.3.46 so the
WASM build and republish run in step with the C++ change.

+++ common/hogql_parser/parser_json.cpp
@@ -327,7 +327,7 @@ class HogQLParseTreeJSONConverter : public HogQLParserBaseVisitor {
         json["node"] = "TupleAccess";
         if (!is_internal) addPositionInfo(json, ctx);
         json["tuple"] = visitAsJSON(ctx->assignmentTarget());
-        json["index"] = stoll(ctx->DECIMAL_LITERAL()->getText());
+        json["index"] = static_cast<int64_t>(stoll(ctx->DECIMAL_LITERAL()->getText()));
         return json;
       }
       // `target . name` — property access, mirrors ColumnExprPropertyAccess.
